requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78